refactor: _constants.js
